Ignore the value of the standalone declaration.
The parser previously raised an error when it encountered a document declaration
stating that the document is not standalone (with standalone=no, meaning that it
depends on an external resource such as a DTD or XSD schema). Both values of
"yes" (the default if not specified) and "no" are now allowed.

The UnsupportedNotStandalone error type is no longer used, and marked as deprecated.

Fixes #37.